Remove YSMenu from frontend site
Too many people are assuming this is correct. It very well is not.

Leave the hacked timebomb bootloader alone as that is still used for
good reasons. Rename m3adapter.com to M3i Zero (GMP-Z003) as only that
cart should be using it, while all other M3s either have YSMenu or Wood.