r4isdhc.com 4.1b kernel
Looks to be an updated and maintained kernel on r4i-sdhc.com's site which doesn't appear to have a timebomb? I set my consoles date to 16/04/2040 and a timebomb did not trigger

I've repacked it as a 7z file with normal compression, zip was too big. Thought it was interesting, I think r4i-sdhc.com's kernels still have a timebomb though