treewide: use deterministic-zip
https://github.com/timo-reymann/deterministic-zip

This solves a logistics issue; zip files always include the modified
timestamp of a file. This doesn't sound bad until you realize that
GitHub Actions always re-clones the entire repository, and the
timestamp is guaranteed to be different.

Because of this, every time we run our Actions workflow, the zip will
have a different checksum, causing git to re-commit *every single file*
to the frontend repository, causing long workflow times due to the
sheer size of each commit.

deterministic-zip hopefully solves this issue by forcing the modified
timestamp to one hardcoded in its binary, so here is to hoping we don't
abuse our git repo any further.