add r4i-d.com; update BL2CK compatibility list
add r4iplat.com and r4i-d.com to the R4DS Pro BL2CK compat. list;
r4i-d.com seems to be related to the R4DS Pro.
The last kernel release (1.41) seems to be broken as games don't load,
whether or not it is just poor game compatibility, I'm not sure.